refactor(executor): extract command execution logic into separate module
- Created src/executor/command.rs with 776 lines
- Extracted execute_and_capture_output() for command substitution
- Extracted execute_single_command() for single command execution
- Extracted execute_pipeline() for pipeline coordination
- Reduced src/executor/mod.rs from ~4,043 to 3,270 lines
- All 554 tests passing
- Part of Phase 1 Step 1.3 of executor refactoring
